+++ src/ui/config/index.ts
@@ -1,0 +1,1 @@
+export const IS_CLEAR_LOGS_VISIBLE = false;